Use computed price ratio as the starting point for price ratio updates (#114)
* Fix tests

* Simplify _computeCurrentFourthRootPriceRatio (#115)

* refactor: save some bytecode; pull the component calculation inside `computeCurrentFourthRootPriceRatio`

* lint

* update snaps

* Relaxes tolerance for approximate equality checks

Updates the tolerance for approximate equality checks in tests.

The previous error tolerance was based on percentages,
which can be difficult to reason about in the context of
ReClamm. The tolerance is changed to a fixed amount of
wei (1 million wei and 100 wei respectively), which provides
a more intuitive and predictable error margin.

---------

Co-authored-by: EndymionJkb <EndymionJkb@gmail.com>