If Out of Range, recalculate the price range using the current price ratio (#113)
* Use the current price ratio to move the pool prices when out of range

* Fix hardhat tests

* Fix test

* Documentation

* Fix E2E tests

* Comment

* Create a test to computePriceRange

* Apply suggestions from code review

Co-authored-by: EndymionJkb <EndymionJkb@gmail.com>

* Fix snaps

* Explain swap exact out

* Refactors fee collection in ReClamm swap test

Simplifies the E2E test by removing price range checks and ensuring all fees from initial pool setup are collected before the primary test logic begins. This focuses the test on swap fees generated during the intended test scenario.

* Use computed price ratio as the starting point for price ratio updates (#114)

* Fix tests

* Simplify _computeCurrentFourthRootPriceRatio (#115)

* refactor: save some bytecode; pull the component calculation inside `computeCurrentFourthRootPriceRatio`

* lint

* update snaps

* Relaxes tolerance for approximate equality checks

Updates the tolerance for approximate equality checks in tests.

The previous error tolerance was based on percentages,
which can be difficult to reason about in the context of
ReClamm. The tolerance is changed to a fixed amount of
wei (1 million wei and 100 wei respectively), which provides
a more intuitive and predictable error margin.

---------

Co-authored-by: EndymionJkb <EndymionJkb@gmail.com>

* Hardhat test moving OOR and price updating

* Fix tests

* Fix version of submodule

* Fix yarn.lock with new version of monorepo

* Refactors price error handling in tests

Updates test error tolerances for price and virtual balance comparisons to reflect expected variance based on different test scenarios.
This allows more accurate and reliable test assertions.

* Fix tests

* Adds virtual balance movement tests

Adds tests to verify virtual balance movements when the pool is out of range
and the price ratio is either concentrating or deconcentrating.
These tests ensure the correctness of virtual balance calculations under various conditions.

* Simplify a bit the tests

* Refactors reClammPool tests for clarity

Improves the structure of the reClammPool tests by introducing a shared beforeEach hook for collecting fees.
This change enhances the readability and maintainability of the tests by reducing code duplication and grouping related tests within describe blocks.

* Uses FixedPoint.ONE for price ratio computation

The computation of the square root of the price ratio uses the `FixedPoint.ONE` constant instead of the raw `1e18` value.

This change ensures consistency and uses the intended fixed-point representation for improved accuracy in calculations.

* Test math function in hardhat

* Check spot prices after swaps to make sure it does not jump

* Apply suggestions from code review

Co-authored-by: Juan Ignacio Ubeira <juani@balancerlabs.dev>

* replace context by describe

* Comment checkSpotPriceAfterSwap

* Add comment to function

* Simplify a bit the swaps and spot price checks in the tests

* Comment swap that puts the pool out of range

* Adjusting Big Error checks

* Fix comment

* Fix tests

* Fix tests

---------

Co-authored-by: EndymionJkb <EndymionJkb@gmail.com>
Co-authored-by: Juan Ignacio Ubeira <juani@balancerlabs.dev>